Start restructuring code
All the logic is being divided in different files. In `bot.ts` a class
that acts as wrapper for the bot is created which exposes the `launch`
and `implement` keywords. The code for commands is progressively being
moved to the `modules` folder and the database code will be moved after
too.